stock_bot/requirements.txt
python-telegram-bot>=21.0
yfinance>=0.2.30
pandas>=2.0
matplotlib>=3.7
